code-documentation-generator/requirements.txt
openai
python-dotenv
pytest
flask